requirements.txt
# --- Core Backend ---
fastapi>=0.120.0
uvicorn>=0.31.0
pydantic>=2.9.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
requests>=2.32.3

# --- OpenAI & Hugging Face APIs ---
openai>=1.52.0
transformers>=4.45.2
datasets>=3.0.2
accelerate>=1.1.1
huggingface-hub>=0.25.2
sentence-transformers>=3.1.0

# --- ML & DL Frameworks ---
torch>=2.5.1
torchvision>=0.20.1
torchaudio>=2.5.1
scikit-learn>=1.5.2
numpy>=1.26.4
pandas>=2.2.3

# --- Computer Vision & OCR ---
Pillow>=10.4.0
opencv-python>=4.10.0.84
pytesseract>=0.3.13

# --- Audio Processing ---
pydub>=0.25.1
soundfile>=0.12.1
librosa>=0.10.2.post1

# --- Retrieval & Embeddings ---
faiss-cpu>=1.8.0
prometheus-client>=0.21.0

# --- Visualization & Reports ---
matplotlib>=3.9.2


# -------prometheus client --------
prometheus>=0.23.1 

# --- Optional Monitoring / DevOps ---
prometheus-fastapi-instrumentator>=7.0.0
grafana-api>=1.0.3


#-----Open AI -----
openai>=0.28.0

# -----tesseract package-----
# install tesseract by below path
#Go to the official Windows installer page:
# https://github.com/UB-Mannheim/tesseract/wiki